scripts/requirements.txt
pandas==2.2.3
pymysql==1.1.1
python-dotenv==1.1.0
scikit_learn==1.6.1
